feat: Add axios package for HTTP requests
This commit adds the `axios` package to the project's dependencies. The `axios` package is a popular library for making HTTP requests in JavaScript. It provides a simple and intuitive API for sending and receiving data from a server.

The purpose of this change is to enhance the functionality of the project by enabling the use of `axios` for making HTTP requests, which can improve the efficiency and reliability of network communication.