chore: Update shasum in snap.manifest.json
This commit updates the shasum value in the snap.manifest.json file. The new shasum value is "tLg3bUgo4x1zeTmvFJ6zeXn7O1/gKiEy5J4QvJIBNNw=".

The purpose of this change is to ensure the integrity of the snap package by updating the shasum value to match the latest version.